Add outlines
