+++ crates/krilla-test/src/main.rs
@@ -43,6 +43,7 @@ mod font;
 mod validate;
 mod image;
 mod mask;
+mod outline;
 
 const REPLACE: Option<&str> = option_env!("REPLACE");
 const STORE: Option<&str> = option_env!("STORE");
